Add an app for folders, including views, models, urls, and tests
This commit adds the first implementation of folders in the project,
which at this point is a copy of how the notes app is organized with
minor changes. Folders and notes are now related through having a
`parent_folder` foreign key in Note's model. Notes app has also
been modified to show the parent folder in the note views. Notes
tests also had to be modified.

I learned about two small details when dealing with the foreign key.
First, when posting a form the foreign key can be given to the form
through the `id` field which Django automatically generates for the
parent folder. Second, `CreateView` automatically generates a drop-down
selector in `NoteCreateView` for the user to choose a parent folder.